+++ src/window/selectwindow.hpp
@@ -39,7 +39,7 @@ namespace Ui
    public:
       void Print(uint32_t line) const;
       void Resize(uint32_t rows, uint32_t columns);
-      void Scroll(int64_t scrollCount);
+      void Scroll(int32_t scrollCount);
       void ScrollTo(uint32_t scrollLine);
 
       uint32_t CurrentLine() const;
